requirements.txt
python-dotenv
crewai
pandas
transformers
pytrends
beautifulsoup4
crewai_tools
pydantic
tweepy
langchain
requests
torch
torchvision
torchaudio